+++ packages/experimental/TabList/src/TabList/TabList.tsx
@@ -8,13 +8,11 @@ import type { UseSlots } from '@fluentui-react-native/framework';
 import { compose, mergeProps, withSlots } from '@fluentui-react-native/framework';
 
 import { stylingSettings } from './TabList.styling';
-import type { TabListType, TabListProps, TabListContextData } from './TabList.types';
+import type { TabListType, TabListProps } from './TabList.types';
 import { tabListName } from './TabList.types';
+import { TabListContext } from './TabListContext';
 import { useTabList } from './useTabList';
-
-export const TabListContext = React.createContext<TabListContextData>({
-  selectedKey: '',
-});
+import TabListAnimatedIndicator from '../TabListAnimatedIndicator/TabListAnimatedIndicator';
 
 export const TabList = compose<TabListType>({
   displayName: tabListName,
@@ -25,23 +23,23 @@ export const TabList = compose<TabListType>({
   },
   useRender: (userProps: TabListProps, useSlots: UseSlots<TabListType>) => {
     // configure props and state for tabs based on user props
-    const tabList = useTabList(userProps);
+    const tablist = useTabList(userProps);
 
     // Grab the styled slots.
     const Slots = useSlots(userProps);
 
     // Return the handler to finish render.
     return (final: TabListProps, ...children: React.ReactNode[]) => {
-      if (!tabList.state) {
+      if (!tablist.state) {
         return null;
       }
 
-      const { disabled, defaultTabbableElement, isCircularNavigation, vertical, ...mergedProps } = mergeProps(tabList.props, final);
+      const { disabled, defaultTabbableElement, isCircularNavigation, vertical, ...mergedProps } = mergeProps(tablist.props, final);
 
       return (
         <TabListContext.Provider
           // Passes in the selected key and a hook function to update the newly selected tab and call the client's onTabsClick callback.
-          value={tabList.state.context}
+          value={tablist.state}
         >
           <Slots.container
             disabled={disabled}
@@ -50,6 +48,7 @@ export const TabList = compose<TabListType>({
             isCircularNavigation={isCircularNavigation}
           >
             <Slots.stack {...mergedProps}>{children}</Slots.stack>
+            {tablist.state.animatedIndicatorStyles && tablist.state.layout && <TabListAnimatedIndicator />}
           </Slots.container>
         </TabListContext.Provider>
       );