+++ packages/experimental/TabList/src/TabList/useTabList.ts
@@ -1,10 +1,12 @@
 import * as React from 'react';
-import type { View, AccessibilityState } from 'react-native';
+import type { View, AccessibilityState, LayoutRectangle } from 'react-native';
 
-import { memoize } from '@fluentui-react-native/framework';
+import { memoize, mergeStyles } from '@fluentui-react-native/framework';
+import type { LayoutEvent } from '@fluentui-react-native/interactive-hooks';
 import { useSelectedKey } from '@fluentui-react-native/interactive-hooks';
 
-import type { TabListProps, TabListInfo, TabListState } from './TabList.types';
+import type { TabListInfo, TabListProps, TabLayoutInfo } from './TabList.types';
+import type { AnimatedIndicatorStyles, AnimatedIndicatorStylesUpdate } from '../TabListAnimatedIndicator/TabListAnimatedIndicator.types';
 
 /**
  * Re-usable hook for TabList.
@@ -34,7 +36,7 @@ export const useTabList = (props: TabListProps): TabListInfo => {
   // selectedTabRef should be set to default tabbable element.
   const [selectedTabRef, setSelectedTabRef] = React.useState(React.useRef<View>(null));
   const [invoked, setInvoked] = React.useState(false);
-  const [tabKeys, setTabKeys] = React.useState([]);
+  const [tabKeys, setTabKeys] = React.useState<string[]>([]);
 
   const addTabKey = React.useCallback(
     (tabKey: string) => {
@@ -53,21 +55,40 @@ export const useTabList = (props: TabListProps): TabListInfo => {
     [setTabKeys],
   );
 
-  const state: TabListState = {
-    context: {
-      addTabKey: addTabKey,
-      appearance: appearance,
-      disabled: disabled,
-      invoked: invoked,
-      onTabSelect: data.onKeySelect,
-      removeTabKey: removeTabKey,
-      selectedKey: selectedKey ?? data.selectedKey,
-      setSelectedTabRef: setSelectedTabRef,
-      setInvoked: setInvoked,
-      size: size,
-      tabKeys: tabKeys,
-      vertical: vertical,
-    },
+  // State variables and functions for saving layout info and other styling information to style the animated indicator.
+  const [canShowAnimatedIndicator, setCanShowAnimatedIndicator] = React.useState<boolean>(false);
+  const [listLayoutMap, setListLayoutMap] = React.useState<{ [key: string]: TabLayoutInfo }>({});
+  const [tabListLayout, setTabListLayout] = React.useState<LayoutRectangle>();
+  const [userDefinedAnimatedIndicatorStyles, setUserDefinedAnimatedIndicatorStyles] = React.useState<AnimatedIndicatorStyles>({
+    container: {},
+    indicator: {},
+  });
+
+  const addTabLayout = (tabKey: string, layoutInfo: TabLayoutInfo) => {
+    setListLayoutMap((prev) => ({ ...prev, [tabKey]: layoutInfo }));
+  };
+
+  const updateStyles = (update: AnimatedIndicatorStylesUpdate) => {
+    if (!update.container && !update.indicator) {
+      return;
+    }
+    setUserDefinedAnimatedIndicatorStyles((prev) => {
+      const newStyles: AnimatedIndicatorStyles = { ...prev };
+      if (update.container) {
+        newStyles.container = mergeStyles(prev.container, update.container);
+      }
+      if (update.indicator) {
+        newStyles.indicator = mergeStyles(prev.indicator, update.indicator);
+      }
+      return newStyles;
+    });
+  };
+
+  // TabList layout callback used to style the animated indicator.
+  const onTabListLayout = (e: LayoutEvent) => {
+    if (e.nativeEvent.layout) {
+      setTabListLayout(e.nativeEvent.layout);
+    }
   };
 
   return {
@@ -80,10 +101,33 @@ export const useTabList = (props: TabListProps): TabListInfo => {
       componentRef: componentRef,
       defaultTabbableElement: selectedTabRef,
       isCircularNavigation: isCircularNavigation ?? false,
+      onLayout: onTabListLayout,
       size: size,
       vertical: vertical,
     },
-    state: { ...state },
+    state: {
+      addTabKey: addTabKey,
+      addTabLayout: addTabLayout,
+      animatedIndicatorStyles: userDefinedAnimatedIndicatorStyles,
+      appearance: appearance,
+      canShowAnimatedIndicator: canShowAnimatedIndicator,
+      disabled: disabled,
+      invoked: invoked,
+      layout: {
+        tablist: tabListLayout,
+        tabs: listLayoutMap,
+      },
+      onTabSelect: data.onKeySelect,
+      removeTabKey: removeTabKey,
+      selectedKey: selectedKey ?? data.selectedKey,
+      setSelectedTabRef: setSelectedTabRef,
+      setCanShowAnimatedIndicator: setCanShowAnimatedIndicator,
+      setInvoked: setInvoked,
+      size: size,
+      tabKeys: tabKeys,
+      vertical: vertical,
+      updateAnimatedIndicatorStyles: updateStyles,
+    },
   };
 };